Split unit and integration tests via filename suffix
Unit tests stay `*.test.ts`, integration tests use `*.integration.test.ts`. Each
package now declares two Vitest projects to run them separately. Adds `test:unit`
and `test:integration` scripts; `yarn test` still runs both. CI now runs unit
tests as a fast early gate, then integration tests as a per-file matrix (each
file gets its own runner + Docker stack) to avoid infra contention. Test
infrastructure only — no API changes.
