+++ package.json
@@ -65,11 +65,14 @@
     "verify": "turbo run verify",
     "verify:check": "turbo run verify:check",
     "verify:test": "turbo run verify:test",
+    "verify:integration": "turbo run verify:integration",
     "publint": "turbo run publint",
     "check-barrel-exports": "node scripts/check-barrel-exports.mjs",
     "typecheck": "turbo run typecheck",
     "lint": "turbo run lint",
     "test": "turbo run test",
+    "test:unit": "turbo run test:unit",
+    "test:integration": "turbo run test:integration",
     "dist": "turbo run dist",
     "watch": "turbo watch dist",
     "dist:publish": "turbo run dist:publish",
