+++ packages/node-client/package.json
@@ -74,6 +74,8 @@
     "gen-test-tx": "DEBUG='testcontainers:containers' tsx ./scripts/generate-test-tx.ts",
     "typecheck": "tsc -b ./tsconfig.json --noEmit --force",
     "test": "DEBUG='testcontainers:containers' vitest run",
+    "test:unit": "vitest run --project unit --passWithNoTests",
+    "test:integration": "DEBUG='testcontainers:containers' vitest run --project integration --passWithNoTests",
     "lint": "eslint --max-warnings 0",
     "format": "prettier --write \"**/*.{ts,js,json,yaml,yml,md}\"",
     "format:check": "prettier --check \"**/*.{ts,js,json,yaml,yml,md}\"",
